move hamiltonian gate
