+++ qiskit/circuit/library/__init__.py
@@ -319,6 +319,15 @@ Particular Quantum Circuits
    GroverOperator
    PhaseOracle
    EvolvedOperatorAnsatz
+
+Time Evolution Gates
+====================
+
+.. autosummary::
+   :toctree: ../stubs/
+   :template: autosummary/class_no_inherited_members.rst
+
+   HamiltonianGate
    PauliEvolutionGate
 
 
@@ -509,8 +518,10 @@ from .generalized_gates import (
     PauliGate,
     LinearFunction,
 )
-
-from .evolution import PauliEvolutionGate
+from .evolution import (
+    HamiltonianGate,
+    PauliEvolutionGate,
+)
 from .boolean_logic import (
     AND,
     OR,
